add more stream functionality
putting all stream stuff in the stream package.  We will deprecate
all of the Iterator and Writer api.  It was too complicated,
this is much, much simpler.  We still need to add toBean method and
migrate new streams for xml, csv, delimited and sql.  We will call
these 'streams' instead of iterators.